backend/pyproject.toml
[tool.poetry]
name = "paystub-service"
version = "0.1.0"
description = "Backend service for paystub uploads and document manifest"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.118.0"
uvicorn = "^0.30.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
pydantic = "^2.6.1"
pydantic-settings = "^2.1.0"
pydantic-ai = "1.4.0"
openai = "^1.0.0"
maxim-py = "^3.5.0"
httpx = "^0.28.0"
beautifulsoup4 = "^4.12.0"
apscheduler = "^3.10.0"
playwright = "^1.49.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.5"
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'

[tool.isort]
profile = "black"
